Add batch and signing tests folders to ristretto255 (#364)
