+++ frost-ristretto255/src/tests.rs
@@ -1,8 +1,10 @@
 #[cfg(test)]
+mod batch_tests;
+#[cfg(test)]
 mod coefficient_commitment_tests;
 #[cfg(test)]
-mod tests;
+mod deserialize_tests;
+#[cfg(test)]
+mod signing_tests;
 #[cfg(test)]
 mod vss_commitment_tests;
-#[cfg(test)]
-mod deserialize_tests;
