+++ frost-ristretto255/src/tests/batch_tests.rs
@@ -1,0 +1,17 @@
+use rand::thread_rng;
+
+use crate::*;
+
+#[test]
+fn check_batch_verify() {
+    let rng = thread_rng();
+
+    frost_core::tests::batch::batch_verify::<Ristretto255Sha512, _>(rng);
+}
+
+#[test]
+fn check_bad_batch_verify() {
+    let rng = thread_rng();
+
+    frost_core::tests::batch::bad_batch_verify::<Ristretto255Sha512, _>(rng);
+}
