+++ frost-ristretto255/src/tests/deserialize_tests.rs
@@ -1,6 +1,6 @@
+use crate::*;
 use curve25519_dalek::{ristretto::RistrettoPoint, traits::Identity};
 use frost_core::{Ciphersuite, Group, GroupError};
-use crate::*;
 
 #[test]
 fn check_deserialize_identity() {
