+++ frost-ristretto255/src/tests/signing_tests.rs
@@ -36,17 +36,3 @@ fn check_sign_with_dealer() {
 
     frost_core::tests::check_sign_with_dealer::<Ristretto255Sha512, _>(rng);
 }
-
-#[test]
-fn check_batch_verify() {
-    let rng = thread_rng();
-
-    frost_core::tests::batch::batch_verify::<Ristretto255Sha512, _>(rng);
-}
-
-#[test]
-fn check_bad_batch_verify() {
-    let rng = thread_rng();
-
-    frost_core::tests::batch::bad_batch_verify::<Ristretto255Sha512, _>(rng);
-}
